# docs/swagger.yaml
basePath: /api
definitions:
...
  model.Star:
    properties:
      age:
        type: string
      description:
        type: string
      distance:
        type: string
      id:
        type: integer
      image_path:
        type: string
      is_deleted:
        type: string
      luminosity:
        type: string
      mass:
        type: string
      metallicity:
        type: string
      radius:
        type: string
      spectral_type:
        type: string
      temperature:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: bitop@spatecon.ru
    name: API Support
    url: https://vk.com/bmstu_schedule
  description: API для работы с экзопланетами и звездами
  license:
    name: AS IS (NO WARRANTY)
  title: ExoCalc API
  version: "1.0"
paths:
...
  /auth/me:
    get:
      consumes:
      - application/json
      description: Возвращает данные авторизованного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получение информации о текущем пользователе
      tags:
      - Auth
...
  /stars:
    get:
      consumes:
      - application/json
      description: Возвращает список всех звезд или отфильтрованный по названию
      parameters:
      - description: Поиск по названию звезды
        in: query
        name: searchedStar
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получение списка звезд
      tags:
      - Stars
    post:
      consumes:
      - application/json
      description: Создает новую звезду (требуется роль Astronomer)
      parameters:
      - description: Данные звезды
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Star'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Star'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создание новой звезды
      tags:
      - Stars
  /stars/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет звезду по ID (требуется роль Astronomer)
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Удаление звезды
      tags:
      - Stars
    get:
      consumes:
      - application/json
      description: Возвращает данные звезды по указанному ID
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Получение звезды по ID
      tags:
      - Stars
    put:
      consumes:
      - application/json
      description: Обновляет данные существующей звезды (требуется роль Astronomer)
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные звезды
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Star'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Star'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Обновление звезды
      tags:
      - Stars
  /stars/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение для звезды (требуется роль Astronomer)
      parameters:
      - description: ID звезды
        in: path
        name: id
        required: true
        type: integer
      - description: Файл изображения
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Загрузка изображения звезды
      tags:
      - Stars
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"